Moved navigation files in one place
