+++ Frontend/App.js
@@ -3,54 +3,15 @@ import {Provider} from 'react-redux';
 import 'react-native-gesture-handler';
 
 import {NavigationContainer} from '@react-navigation/native';
-import {createStackNavigator} from '@react-navigation/stack';
 
-import SplashScreen from './src/screens/splashScreen';
-import LoginScreen from './src/screens/loginScreen';
-import RegisterScreen from './src/screens/registerScreen';
-import TabNavigatorRoutes from './src/screens/tabNavigatorRoutes';
+import AppNavigtorRoutes from './src/navigation/stackNavigatorRotues';
 import store from './src/redux/store';
-
-const Stack = createStackNavigator();
-
-const Auth = () => {
-  return (
-    <Stack.Navigator initialRouteName="LoginScreen">
-      <Stack.Screen
-        name="LoginScreen"
-        component={LoginScreen}
-        options={{headerShown: false}}
-      />
-      <Stack.Screen
-        name="RegisterScreen"
-        component={RegisterScreen}
-        options={{headerShown: false}}
-      />
-    </Stack.Navigator>
-  );
-};
 
 const App = () => {
   return (
     <Provider store={store}>
       <NavigationContainer>
-        <Stack.Navigator initialRouteName="SplashScreen">
-          <Stack.Screen
-            name="SplashScreen"
-            component={SplashScreen}
-            options={{headerShown: false}}
-          />
-          <Stack.Screen
-            name="Auth"
-            component={Auth}
-            options={{headerShown: false}}
-          />
-          <Stack.Screen
-            name="TabNavigatorRoutes"
-            component={TabNavigatorRoutes}
-            options={{headerShown: false}}
-          />
-        </Stack.Navigator>
+        <AppNavigtorRoutes />
       </NavigationContainer>
     </Provider>
   );
